backend_FastAPI/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
psycopg2-binary
python-dotenv
ortools
pydantic
passlib
python-jose
passlib[bcrypt]
pandas
python-multipart
openpyxl
xlrd